requirements.txt
openai>=1.30.0
anthropic>=0.25.0
google-genai>=1.0.0
tqdm>=4.65.0
